dev/local 패키지 분리 및 debug tool install, dev db create/psql

+++ app/tactical_dolls/models/doll.py
@@ -29,13 +29,13 @@ class Doll(models.Model):
         ('rf', 'rf'),
     )
 
-    id = models.PositiveSmallIntegerField(unique=True, primary_key=True, verbose_name='전술 인형 넘버링')
+    id = models.PositiveIntegerField(unique=True, primary_key=True, verbose_name='전술 인형 넘버링')
     codename = models.CharField(max_length=50, blank=True, null=True, verbose_name='전술 인형 명칭')
     kr_name = models.CharField(max_length=50, blank=True, null=True)
     rank = models.CharField(max_length=5, choices=RANK_CHOICE, verbose_name='전술 인형 등급')
     type = models.CharField(max_length=3, choices=TYPE_CHOICE, verbose_name='전술 인형 타입')
-    build_time = models.PositiveSmallIntegerField(blank=True, null=True, verbose_name='전술 인형 제조 시간')
-    grow = models.PositiveSmallIntegerField(blank=True, null=True)
+    build_time = models.PositiveIntegerField(blank=True, null=True, verbose_name='전술 인형 제조 시간')
+    grow = models.PositiveIntegerField(blank=True, null=True)
     image = models.ImageField(upload_to='doll_image', blank=True, null=True)
     image_d = models.ImageField(upload_to='doll_image', blank=True, null=True)
     equip_slot_01 = models.CharField(max_length=100, blank=True, null=True)
